# api/docs/swagger.yaml
definitions:
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.Client:
    properties:
      active:
        type: boolean
      birthday:
        type: string
      branch_id:
        type: string
      created_at:
        type: string
      father_name:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
  models.Coming:
    properties:
      branch_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      increment_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.CreateClient:
    properties:
      active:
        type: boolean
      birthday:
        type: string
      branch_id:
        type: string
      father_name:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  models.CreateComing:
    properties:
      branch_id:
        type: string
      date:
        type: string
      increment_id:
        type: string
    type: object
...
  models.UpdateClient:
    properties:
      active:
        type: boolean
      birthday:
        type: string
      branch_id:
        type: string
      father_name:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      last_name:
        type: string
      phone:
        type: string
    type: object
  models.UpdateComing:
    properties:
      branch_id:
        type: string
      date:
        type: string
      id:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /client:
    get:
      consumes:
      - application/json
      description: Get List Client details by its ok.
      parameters:
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client details
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Client not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get List Client
      tags:
      - Client
    post:
      consumes:
      - application/json
      description: Create Client
      parameters:
      - description: Client ID
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.CreateClient'
      produces:
      - application/json
      responses:
        "200":
          description: Client details
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Client not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: create a Client
      tags:
      - Client
  /client/{id}:
    delete:
      consumes:
      - application/json
      description: Get List Client details by its ok
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client details
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Client not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get List Client
      tags:
      - Client
    get:
      consumes:
      - application/json
      description: Get Client details by its ID.
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client details
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Client not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a Client by ID
      tags:
      - Client
    put:
      consumes:
      - application/json
      description: Get List Client details by its ok.
      parameters:
      - description: models.UpdateClient
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/models.UpdateClient'
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client details
          schema:
            $ref: '#/definitions/models.Client'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Client not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update Client
      tags:
      - Client
  /coming:
    get:
      consumes:
      - application/json
      description: Get a list of Comings with optional filtering.
      parameters:
      - description: Number of items to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of Comings
          schema:
            items:
              $ref: '#/definitions/models.Coming'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get a list of Comings
      tags:
      - Coming
    post:
      consumes:
      - application/json
      description: Create a new Coming in the market system.
      parameters:
      - description: Coming information
        in: body
        name: Coming
        required: true
        schema:
          $ref: '#/definitions/models.CreateComing'
      produces:
      - application/json
      responses:
        "201":
          description: Created Coming
          schema:
            $ref: '#/definitions/models.Coming'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create a new Coming
      tags:
      - Coming
  /coming/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing Coming.
      parameters:
      - description: Coming ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete an Coming
      tags:
      - Coming
    get:
      consumes:
      - application/json
      description: Get Coming details by its ID.
      parameters:
      - description: Coming ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Coming details
          schema:
            $ref: '#/definitions/models.Coming'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Coming not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Get an Coming by ID
      tags:
      - Coming
    put:
      consumes:
      - application/json
      description: Update an existing Coming.
      parameters:
      - description: Coming ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Coming information
        in: body
        name: Coming
        required: true
        schema:
          $ref: '#/definitions/models.UpdateComing'
      produces:
      - application/json
      responses:
        "202":
          description: Updated Coming
          schema:
            $ref: '#/definitions/models.Coming'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Coming not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Update an Coming
      tags:
      - Coming
...
swagger: "2.0"